+++ app.js
@@ -5,8 +5,8 @@ const cookieParser = require('cookie-parser')
 const logger = require('morgan')
 const cors = require('cors')
 require('dotenv').config()
+const routerApi = require('./routes/index');
 
-const indexRouter = require('./routes/index')
 
 const port = process.env.PORT || 3000
 
@@ -19,7 +19,8 @@ app.use(express.urlencoded({ extended: false }))
 app.use(cookieParser())
 app.use(express.static(path.join(__dirname, 'public')))
 
-app.use('/', indexRouter)
+
+routerApi(app);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
