+++ routes/accounts/index.js
@@ -1,0 +1,10 @@
+const express = require('express')
+
+const router = express.Router()
+
+router.get('/', async (req, res) => {
+   const rta = 'Accounts';
+   res.send(rta);
+ })
+
+module.exports = router
