+++ routes/users/index.js
@@ -1,0 +1,10 @@
+const express = require('express')
+
+const router = express.Router()
+
+router.get('/', async (req, res) => {
+   const rta = 'hola mundo';
+   res.send(rta);
+ })
+
+module.exports = router
